Fix: Beautyinfo
