+++ src/main/java/org/example/sivillage/member/domain/ScalpTone.java
@@ -1,4 +1,4 @@
-package org.example.sivillage.memberinfo.domain;
+package org.example.sivillage.member.domain;
 
 import lombok.RequiredArgsConstructor;
