+++ src/main/java/org/example/sivillage/member/infrastructure/MemberRepository.java
@@ -8,6 +8,5 @@ import java.util.Optional;
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
     Boolean existsByEmail(String email);
-
     Optional<Member> findByMemberUuid(String memberUuid);
 }
