+++ src/main/java/org/example/sivillage/member/infrastructure/SizeInfoRepository.java
@@ -1,6 +1,6 @@
-package org.example.sivillage.memberinfo.infrastructure;
+package org.example.sivillage.member.infrastructure.infrastructure;
 
-import org.example.sivillage.memberinfo.domain.SizeInfo;
+import org.example.sivillage.member.domain.SizeInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SizeInfoRepository extends JpaRepository<SizeInfo, Long>{
