--- src/main/java/org/example/sivillage/memberinfo/infrastructure/BeautyInfoRepository.java
@@ -1,8 +1,0 @@
-package org.example.sivillage.memberinfo.infrastructure;
-
-import org.example.sivillage.memberinfo.domain.BeautyInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BeautyInfoRepository extends JpaRepository<BeautyInfo, Long> {
-
-}
